Reversed the set of modules in the ROM.
Reading documentation about the Acorn ROM module system, I spotted the
module data structure matched the ones in the ROM, so I used this to
identify the modules. I have not gone as far as identifying the
routines for the individual system calls (trying not to geek-snipe
myself further), but it should be relatively straightforward to do so.